third_party/WebKit/LayoutTests/fast/animation/scroll-animations/scrolltimeline-root-scroller.html

<!DOCTYPE html>
<style>
html {
  /* Ensure the document is scrollable. */
  min-height: 100%;
  min-width: 100%;
  padding-bottom: 100px;
  padding-right: 100px;
}
</style>

<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>

<script>
test(function() {
  const scroller = document.scrollingElement;
  // For simplicity, we set the timeRange(s) such that currentTime maps directly to
  // the value scrolled.
  const scrollerHeight = scroller.scrollHeight - scroller.clientHeight;
  const scrollerWidth = scroller.scrollWidth - scroller.clientWidth;

  const blockScrollTimeline = new ScrollTimeline(
      { scrollSource: scroller, timeRange: scrollerHeight, orientation: 'block' });
  const inlineScrollTimeline = new ScrollTimeline(
      { scrollSource: scroller, timeRange: scrollerWidth, orientation: 'inline' });

  // Unscrolled, both timelines should read a currentTime of 0.
  assert_equals(blockScrollTimeline.currentTime, 0);
  assert_equals(inlineScrollTimeline.currentTime, 0);

  // Now do some scrolling and make sure that the ScrollTimelines update.
  scroller.scrollTop = 50;
  scroller.scrollLeft = 75;

  // As noted above, we have mapped timeRange such that currentTime should just
  // be the scroll offset.
  assert_equals(blockScrollTimeline.currentTime, 50);
  assert_equals(inlineScrollTimeline.currentTime, 75);
}, 'currentTime calculates the correct time for a document.scrollingElement scrollSource');
</script>
